fix fallthrough bug in game loop
somehow, I managed to insert the case statement for restarting the game in
between the cases for flagging with the mouse and keyboard, which are connected
together with a fallthrough. That made the game restart whenever the mouse was
used to flag. I didn't notice this earlier, since I mostly play this game with the keyboard. (shame on me for not testing my code :| )